requirements-lock.txt
alembic==1.9.1
alibi-detect==0.10.4
attrs==22.2.0
autopage==0.5.1
blosc2==2.0.0
catalogue==2.0.8
charset-normalizer==2.0.12
cliff==4.1.0
cmaes==0.9.0
cmd2==2.4.2
colorlog==6.7.0
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.32
dataclasses==0.6
dill==0.3.6
evidently==0.2.1
filelock==3.9.0
fonttools==4.38.0
frouros==0.1.0
greenlet==2.0.1
h5py==3.7.0
huggingface-hub==0.11.1
idna==3.4
imageio==2.23.0
imbalanced-learn==0.10.1
importlib-metadata==4.13.0
importlib-resources==5.10.2
joblib==1.2.0
kiwisolver==1.4.4
llvmlite==0.36.0
Mako==1.2.4
MarkupSafe==2.1.1
matplotlib==3.6.2
menelaus==0.1.2
mergedeep==1.3.4
msgpack==1.0.4
networkx==2.8.8
numba==0.53.1
numexpr==2.8.4
numpy==1.23.5
opencv-python==4.7.0.68
optuna==3.0.5
packaging==22.0
pandas==1.4.1
patsy==0.5.3
pbr==5.11.0
Pillow==9.4.0
plotly==5.11.0
prettytable==3.5.0
py-cpuinfo==9.0.0
pydantic==1.10.4
pyparsing==3.0.9
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2022.7
PyWavelets==1.4.1
PyYAML==5.4.1
regex==2022.10.31
requests==2.27.1
river==0.14.0
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.8.1
seaborn==0.12.2
serval-config-utils==0.1.4
serval-ml-commons==0.1.4 
setuptools-scm==7.1.0
six==1.16.0
sklearn==0.0.post1
SQLAlchemy==1.4.45
statsmodels==0.13.5
stevedore==4.1.1
tables==3.8.0
tenacity==8.1.0
threadpoolctl==3.1.0
tifffile==2022.10.10
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
torch==1.12.1
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.13
uuid==1.30
wcwidth==0.2.5
zipp==3.11.0
